requirements.txt
aioredis==1.2.0
amqp==2.3.2
ansible==2.6.6
anyjson==0.3.3
asgi-redis==1.4.3
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.10.1
bcrypt==3.1.4
billiard==3.5.0.4
celery==4.2.1
celery-with-redis==3.0
certifi==2018.10.15
cffi==1.11.5
channels==2.1.6
channels-redis==2.2.1
chardet==3.0.4
configparser==3.5.0
constantly==15.1.0
cryptography==2.3.1
daphne==2.2.2
DBUtils==1.3
Django==2.1.4
django-auth-ldap==1.5.0
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-db==0.0.7
django-timezone-field==3.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
future==0.15.2
hiredis==1.0.0
httplib2==0.9.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
jdcal==1.4
Jinja2==2.10
kombu==4.2.1
ldap==1.0.2
ldap3==2.5.1
MarkupSafe==1.1.0
msgpack==0.5.6
msgpack-python==0.5.6
mysql-replication==0.18
mysqlclient==1.3.13
numpy==1.15.4
odfpy==1.3.6
openpyxl==2.5.3
pandas==0.23.4
paramiko==2.4.1
Pillow==5.0.0
ply==3.11
pyasn1==0.4.2
pyasn1-modules==0.2.3
pycparser==2.18
pycrypto==2.6.1
pycryptodomex==3.6.1
PyHamcrest==1.9.0
PyJWT==1.6.4
PyMySQL==0.9.2
PyNaCl==1.2.1
python-crontab==2.3.5
python-ldap==3.0.0
python-magic==0.4.15
pytz==2018.7
PyYAML==3.13
redis==2.10.6
requests==2.20.0
six==1.11.0
tablib==0.12.1
Twisted==18.9.0
txaio==18.8.1
unicodecsv==0.14.1
urllib3==1.24.1
vine==1.1.4
webencodings==0.5.1
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.6.0
